requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
py-clob-client>=0.13.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
